Add files via upload
CURRENCY CONVERTOR :

The purpose of this Java project is to convert a currency from one to another. We see variations with different countries using different currencies. Be aware of the current exchange prices in the market and then can convert from one currency to another. A calculator-like application, developed using Ajax, Applet, and web features of Java servlets. You also get a regular update when there’s a change in the value of the country’s currency and also the conversion rate.